Fix bad symlink creation in install script, fix check to see if files/symlinks already exist.

+++ install.sh
@@ -3,17 +3,20 @@
 # Colors
 RED='\033[0;31m'
 GREEN='\033[0;32m'
+BLUE='\033[0;34m'
 RESET='\033[0m'
 
-echo Check if configs already exist for files that will be symlinked.
+echo -e "${BLUE}Check if configs already exist for files that will be symlinked.${RESET}"
 
-if [[ -e /etc/nixos/configuration.nix ]]; then
+# Check if the file exists (-e) and that it is not a symlink (-L)
+# Only backup files, not symlinks
+if [[ -e /etc/nixos/configuration.nix && ! -L /etc/nixos/configuration.nix ]]; then
   echo /etc/nixos/configuration.nix exists and needs to be removed so a symlink can be created.
   if [[ -w /etc/nixos/configuration.nix ]]; then
     echo /etc/nixos/configuration.nix is writable, create backup
     mv /etc/nixos/configuration.nix /etc/nixos/configuration.nix.bak
     if [ $? -eq 0 ]; then
-      echo -e "${GREEN}/etc/nixos/configuration.nix backup created${RESET}"
+      echo -e "${GREEN}/etc/nixos/configuration.nix removed, /etc/nixos/configuration.nix.bak created${RESET}"
     else
       echo -e "${RED}/etc/nixos/configuration.nix backup failed${RESET}"
       exit 1
@@ -26,13 +29,15 @@ else
   echo -e "${GREEN}/etc/nixos/configuration.nix does not exist${RESET}"
 fi
 
-if [[ -e /etc/nixos/nvidia-drivers.nix ]]; then
+# Check if the file exists (-e) and that it is not a symlink (-L)
+# Only backup files, not symlinks
+if [[ -e /etc/nixos/nvidia-drivers.nix && ! -L /etc/nixos/configuration.nix ]]; then
   echo /etc/nixos/nvidia-drivers.nix exists and needs to be removed so a symlink can be created.
   if [[ -w /etc/nixos/nvidia-drivers.nix ]]; then
     echo /etc/nixos/nvidia-drivers.nix is writable, create backup
     mv /etc/nixos/nvidia-drivers.nix /etc/nixos/nvidia-drivers.nix.bak
     if [ $? -eq 0 ]; then
-      echo -e "${GREEN}/etc/nixos/nvidia-drivers.nix backup created${RESET}"
+      echo -e "${GREEN}/etc/nixos/nvidia-drivers.nix removed, /etc/nixos/nvidia-drivers.nix.bak created${RESET}"
     else
       echo -e "${RED}/etc/nixos/nvidia-drivers.nix backup failed${RESET}"
       exit 1
@@ -45,9 +50,22 @@ else
   echo -e "${GREEN}/etc/nixos/nvidia-drivers.nix does not exist${RESET}"
 fi
 
-echo Create symlinks for configs from git
+echo -e "${BLUE}Create symlinks for configs from git${RESET}"
+CURRENT_DIR=$(pwd)
 
-ln -s configuration.nix /etc/nixos/configuration.nix
+if [[ -L /etc/nixos/configuration.nix ]]; then
+  echo Symlink already exists for /etc/nixos/configuration.nix
+  echo Attempt to remove the existing symlink to be replaced with our symlink
+  rm /etc/nixos/configuration.nix
+  if [ $? -eq 0 ]; then
+    echo -e "${GREEN}Existing /etc/nixos/configuration.nix symlink removed${RESET}"
+  else
+    echo -e "${RED}Unable to remove existing /etc/nixos/configuration.nix symlink${RESET}"
+    exit 1
+  fi
+fi
+
+ln -s "${CURRENT_DIR}/nixos/configuration.nix" /etc/nixos/configuration.nix
 if [ $? -eq 0 ]; then
   echo -e "${GREEN}/etc/nixos/configuration.nix symlink created${RESET}"
 else
@@ -55,7 +73,19 @@ else
   exit 1
 fi
 
-ln -s nvidia-drivers.nix /etc/nixos/nvidia-drivers.nix
+if [[ -L /etc/nixos/nvidia-drivers.nix ]]; then
+  echo Symlink already exists for /etc/nixos/nvidia-drivers.nix
+  echo Attempt to remove the existing symlink to be replaced with our symlink
+  rm /etc/nixos/nvidia-drivers.nix
+  if [ $? -eq 0 ]; then
+    echo -e "${GREEN}Existing /etc/nixos/nvidia-drivers.nix symlink removed${RESET}"
+  else
+    echo -e "${RED}Unable to remove existing /etc/nixos/nvidia-drivers.nix symlink${RESET}"
+    exit 1
+  fi
+fi
+
+ln -s "${CURRENT_DIR}/nixos/nvidia-drivers.nix" /etc/nixos/nvidia-drivers.nix
 if [ $? -eq 0 ]; then
   echo -e "${GREEN}/etc/nixos/nvidia-drivers.nix symlink created${RESET}"
 else
